feat: separate text_rank function to text_rank.py
